v2.0.0-dev.2 Add template admin REST APIs
Also refactored the normalisation of image attr strings to be more
robust. Strings are now converted to lower case inside ImageAttrs
instead of requiring callers to know and do that first. This simplified
template loading and validation, and means 20 fewer lines in views.py
and a cleaner constructor for ImageAttrs.