Test users should belong to the "everyone" group
This brings the test users in line with normally created users, and is
also required for portfolios. The affected tests are getting a bit
messy and would benefit from some refactoring (e.g. using `setUp` to
reset permissions, removing hard coded references to the `Red Dwarf`
group) but this will have to wait for another time.